refactor: mv attachedobject attached_object
