+++ pkg/api-server/v0/docs/swagger.json
@@ -14490,29 +14490,22 @@
         "v0.ObjectType": {
             "type": "string",
             "enum": [
-                "AwsAccount",
-                "AwsEksKubernetesRuntimeDefinition",
-                "AwsEksKubernetesRuntimeInstance",
-                "AwsRelationalDatabaseDefinition",
-                "AwsRelationalDatabaseInstance",
-                "AwsObjectStorageBucketDefinition",
-                "AwsObjectStorageBucketInstance",
-                "SecretDefinition",
-                "SecretInstance",
-                "LogBackend",
-                "LogStorageDefinition",
-                "LogStorageInstance",
+                "GatewayDefinition",
+                "GatewayInstance",
+                "GatewayHttpPort",
+                "GatewayTcpPort",
+                "DomainNameDefinition",
+                "DomainNameInstance",
                 "AttachedObjectReference",
-                "Profile",
-                "Tier",
-                "TerraformDefinition",
-                "TerraformInstance",
-                "HelmWorkloadDefinition",
-                "HelmWorkloadInstance",
                 "KubernetesRuntimeDefinition",
                 "KubernetesRuntimeInstance",
-                "ControlPlaneDefinition",
-                "ControlPlaneInstance",
+                "TerraformDefinition",
+                "TerraformInstance",
+                "Profile",
+                "Tier",
+                "HelmWorkloadDefinition",
+                "HelmWorkloadInstance",
+                "AttachedObjectReference",
                 "ObservabilityStackDefinition",
                 "ObservabilityStackInstance",
                 "ObservabilityDashboardDefinition",
@@ -14521,43 +14514,43 @@
                 "MetricsInstance",
                 "LoggingDefinition",
                 "LoggingInstance",
-                "GatewayDefinition",
-                "GatewayInstance",
-                "GatewayHttpPort",
-                "GatewayTcpPort",
-                "DomainNameDefinition",
-                "DomainNameInstance",
-                "AttachedObjectReference",
+                "LogBackend",
+                "LogStorageDefinition",
+                "LogStorageInstance",
+                "SecretDefinition",
+                "SecretInstance",
                 "WorkloadDefinition",
                 "WorkloadResourceDefinition",
                 "WorkloadInstance",
                 "WorkloadResourceInstance",
-                "WorkloadEvent"
+                "WorkloadEvent",
+                "AwsAccount",
+                "AwsEksKubernetesRuntimeDefinition",
+                "AwsEksKubernetesRuntimeInstance",
+                "AwsRelationalDatabaseDefinition",
+                "AwsRelationalDatabaseInstance",
+                "AwsObjectStorageBucketDefinition",
+                "AwsObjectStorageBucketInstance",
+                "ControlPlaneDefinition",
+                "ControlPlaneInstance"
             ],
             "x-enum-varnames": [
-                "ObjectTypeAwsAccount",
-                "ObjectTypeAwsEksKubernetesRuntimeDefinition",
-                "ObjectTypeAwsEksKubernetesRuntimeInstance",
-                "ObjectTypeAwsRelationalDatabaseDefinition",
-                "ObjectTypeAwsRelationalDatabaseInstance",
-                "ObjectTypeAwsObjectStorageBucketDefinition",
-                "ObjectTypeAwsObjectStorageBucketInstance",
-                "ObjectTypeSecretDefinition",
-                "ObjectTypeSecretInstance",
-                "ObjectTypeLogBackend",
-                "ObjectTypeLogStorageDefinition",
-                "ObjectTypeLogStorageInstance",
+                "ObjectTypeGatewayDefinition",
+                "ObjectTypeGatewayInstance",
+                "ObjectTypeGatewayHttpPort",
+                "ObjectTypeGatewayTcpPort",
+                "ObjectTypeDomainNameDefinition",
+                "ObjectTypeDomainNameInstance",
                 "ObjectTypeAttachedObjectReference",
-                "ObjectTypeProfile",
-                "ObjectTypeTier",
-                "ObjectTypeTerraformDefinition",
-                "ObjectTypeTerraformInstance",
-                "ObjectTypeHelmWorkloadDefinition",
-                "ObjectTypeHelmWorkloadInstance",
                 "ObjectTypeKubernetesRuntimeDefinition",
                 "ObjectTypeKubernetesRuntimeInstance",
-                "ObjectTypeControlPlaneDefinition",
-                "ObjectTypeControlPlaneInstance",
+                "ObjectTypeTerraformDefinition",
+                "ObjectTypeTerraformInstance",
+                "ObjectTypeProfile",
+                "ObjectTypeTier",
+                "ObjectTypeHelmWorkloadDefinition",
+                "ObjectTypeHelmWorkloadInstance",
+                "ObjectTypeAttachedObjectReference",
                 "ObjectTypeObservabilityStackDefinition",
                 "ObjectTypeObservabilityStackInstance",
                 "ObjectTypeObservabilityDashboardDefinition",
@@ -14566,18 +14559,25 @@
                 "ObjectTypeMetricsInstance",
                 "ObjectTypeLoggingDefinition",
                 "ObjectTypeLoggingInstance",
-                "ObjectTypeGatewayDefinition",
-                "ObjectTypeGatewayInstance",
-                "ObjectTypeGatewayHttpPort",
-                "ObjectTypeGatewayTcpPort",
-                "ObjectTypeDomainNameDefinition",
-                "ObjectTypeDomainNameInstance",
-                "ObjectTypeAttachedObjectReference",
+                "ObjectTypeLogBackend",
+                "ObjectTypeLogStorageDefinition",
+                "ObjectTypeLogStorageInstance",
+                "ObjectTypeSecretDefinition",
+                "ObjectTypeSecretInstance",
                 "ObjectTypeWorkloadDefinition",
                 "ObjectTypeWorkloadResourceDefinition",
                 "ObjectTypeWorkloadInstance",
                 "ObjectTypeWorkloadResourceInstance",
-                "ObjectTypeWorkloadEvent"
+                "ObjectTypeWorkloadEvent",
+                "ObjectTypeAwsAccount",
+                "ObjectTypeAwsEksKubernetesRuntimeDefinition",
+                "ObjectTypeAwsEksKubernetesRuntimeInstance",
+                "ObjectTypeAwsRelationalDatabaseDefinition",
+                "ObjectTypeAwsRelationalDatabaseInstance",
+                "ObjectTypeAwsObjectStorageBucketDefinition",
+                "ObjectTypeAwsObjectStorageBucketInstance",
+                "ObjectTypeControlPlaneDefinition",
+                "ObjectTypeControlPlaneInstance"
             ]
         },
         "v0.ObservabilityDashboardDefinition": {
