+++ pkg/api-server/v0/handlers/attached_object_gen.go
