+++ pkg/api-server/v0/routes/attached_object_gen.go
