+++ pkg/api-server/v0/versions/attached_object_gen.go
