+++ pkg/api/v0/attached_object.go
