+++ pkg/client/v0/attached_object_gen.go
