+++ pkg/client/v0/workload.go
@@ -202,75 +202,6 @@ func DeleteWorkloadEventsByQueryString(apiClient *http.Client, apiAddr string, q
 	return &workloadEvents, nil
 }
 
-// GetAttachedObjectReferencesByObjectID fetches an attached object reference
-// by object ID.
-func GetAttachedObjectReferencesByObjectID(apiClient *http.Client, apiAddr string, id uint) (*[]v0.AttachedObjectReference, error) {
-	var attachedObjectReferences []v0.AttachedObjectReference
-
-	response, err := GetResponse(
-		apiClient,
-		fmt.Sprintf("%s%s?objectid=%d", apiAddr, v0.PathAttachedObjectReferences, id),
-		http.MethodGet,
-		new(bytes.Buffer),
-		map[string]string{},
-		http.StatusOK,
-	)
-	if err != nil {
-		return &attachedObjectReferences, fmt.Errorf("call to threeport API returned unexpected response: %w", err)
-	}
-
-	jsonData, err := json.Marshal(response.Data)
-	if err != nil {
-		return &attachedObjectReferences, fmt.Errorf("failed to marshal response data from threeport API: %w", err)
-	}
-
-	decoder := json.NewDecoder(bytes.NewReader(jsonData))
-	decoder.UseNumber()
-	if err := decoder.Decode(&attachedObjectReferences); err != nil {
-		return nil, fmt.Errorf("failed to decode object in response data from threeport API: %w", err)
-	}
-
-	return &attachedObjectReferences, nil
-}
-
-// EnsureAttachedObjectReferenceExists ensures that an attached object reference
-// exists for the given object type and ID.
-func EnsureAttachedObjectReferenceExists(
-	apiClient *http.Client,
-	apiAddr string,
-	objectType string,
-	objectID *uint,
-	attachedObjectType string,
-	attachedObjectID *uint,
-) error {
-	attachedObjectReferences, err := GetAttachedObjectReferencesByObjectID(apiClient, apiAddr, *objectID)
-	if err != nil {
-		return fmt.Errorf("failed to get attached object references by object ID: %w", err)
-	}
-
-	// check if attached object reference already exists
-	for _, attachedObjectReference := range *attachedObjectReferences {
-		if *attachedObjectReference.AttachedObjectType == attachedObjectType &&
-			*attachedObjectReference.AttachedObjectID == *attachedObjectID {
-			return nil
-		}
-	}
-
-	// create attached object reference
-	workloadInstanceAttachedObjectReference := &v0.AttachedObjectReference{
-		ObjectID:           objectID,
-		ObjectType:         &objectType,
-		AttachedObjectType: &attachedObjectType,
-		AttachedObjectID:   attachedObjectID,
-	}
-	_, err = CreateAttachedObjectReference(apiClient, apiAddr, workloadInstanceAttachedObjectReference)
-	if err != nil {
-		return fmt.Errorf("failed to create attached object reference: %w", err)
-	}
-
-	return nil
-}
-
 // ConfirmWorkloadInstanceReconciled confirms whether a workload instance
 // is reconciled.
 func ConfirmWorkloadInstanceReconciled(
